refactor: Console 애플리케이션을 분리,   service를 제외하고 스프링 컨포넌트화 해제, 콘솔 데이터 저장 방식 변경
