+++ src/main/java/racingcar/RacingCarWebApplication.java
@@ -1,0 +1,12 @@
+package racingcar;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class RacingCarWebApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(RacingCarWebApplication.class, args);
+	}
+}
