+++ src/test/java/racingcar/RacingCarWebApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class RacingCarApplicationTests {
+class RacingCarWebApplicationTests {
 
 	@Test
 	void contextLoads() {
